+++ webio-macros/src/lib.rs
@@ -4,9 +4,10 @@ use quote::{quote, ToTokens};
 use syn::{
     parse::{Parse, ParseStream},
     parse_macro_input,
-    token::{Comma, Semi},
+    token,
     Expr,
     ItemFn,
+    Pat,
 };
 
 struct JoinInput {
@@ -15,7 +16,8 @@ struct JoinInput {
 
 impl Parse for JoinInput {
     fn parse(input: ParseStream) -> syn::Result<Self> {
-        let futures = input.parse_terminated::<Expr, Comma>(Expr::parse)?;
+        let futures =
+            input.parse_terminated::<Expr, token::Comma>(Expr::parse)?;
         Ok(Self { futures: futures.into_iter().collect() })
     }
 }
@@ -23,7 +25,15 @@ impl Parse for JoinInput {
 /// Joins a list of futures, and returns their output into a tuple in the same
 /// order that the futures were given. However, the future must be `'static`.
 ///
-/// # Example
+/// Syntax:
+///
+/// ```ignore
+/// join!(future0, future1, future2, ..., future_n)
+/// ```
+///
+/// # Examples
+///
+/// ## With Timeout
 /// ```ignore
 /// use std::time::Duration;
 /// use webio::{join, task, time::timeout};
@@ -136,6 +146,168 @@ pub fn join(raw_input: TokenStream) -> TokenStream {
     expanded.into()
 }
 
+struct SelectArm {
+    pattern: Pat,
+    future: Expr,
+    output: Expr,
+}
+
+impl Parse for SelectArm {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let pattern = input.parse()?;
+        input.parse::<token::Eq>()?;
+        let future = input.parse()?;
+        input.parse::<token::FatArrow>()?;
+        let output = input.parse()?;
+        Ok(Self { pattern, future, output })
+    }
+}
+
+struct SelectInput {
+    arms: Vec<SelectArm>,
+}
+
+impl Parse for SelectInput {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let arms = input
+            .parse_terminated::<SelectArm, token::Comma>(SelectArm::parse)?;
+        Ok(Self { arms: arms.into_iter().collect() })
+    }
+}
+
+/// Listens to a list of futures and finishes when the first future finishes,
+/// which is then selected. Every future is placed in a "match arm", and when it
+/// is selected, the "arm" pattern is matched, and the macro evaluates to the
+/// right side of the "arm". The pattern must be irrefutable, typically just
+/// a variable name, or destructuring.
+///
+/// Syntax:
+///
+/// ```ignore
+/// select! {
+///     pattern0 = future0 => output0,
+///     pattern1 = future1 => output1,
+///     pattern2 = future2 => output2,
+///     ...,
+///     pattern_n = future_n => output_n,
+/// }
+/// ```
+///
+/// # Examples
+///
+/// ## With Timeout
+///
+/// ```ignore
+/// use webio::{task, select, time::timeout};
+/// use std::time::Duration;
+///
+/// # fn main () {
+/// # task::detach(async {
+/// let first_handle = task::spawn(async {
+///     timeout(Duration::from_millis(500)).await;
+///     3u32
+/// });
+/// let second_handle = task::spawn(async {
+///     timeout(Duration::from_millis(50)).await;
+///     5u32
+/// });
+/// let third_handle = task::spawn(async {
+///     timeout(Duration::from_millis(350)).await;
+///     7u32
+/// });
+/// let output = select! {
+///     first_val = first_handle => first_val.unwrap() + 10,
+///     second_val = second_handle => second_val.unwrap() + 20,
+///     third_val = third_handle => third_val.unwrap() - 5
+/// };
+/// assert_eq!(output, 25);
+/// # });
+/// # }
+/// ```
+#[proc_macro]
+pub fn select(raw_input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(raw_input as SelectInput);
+    let arms = input.arms;
+
+    let future_var_names = || {
+        (0 .. arms.len())
+            .map(|i| Ident::new(&format!("future{}", i), Span::mixed_site()))
+    };
+
+    let future_decls = future_var_names()
+        .zip(arms.iter().map(|arm| &arm.future))
+        .map(|(ident, future)| {
+            quote! { let #ident = #future; }
+        });
+
+    let output_var_name = Ident::new("output", Span::mixed_site());
+
+    let output_decl = quote! {
+        let #output_var_name= ::std::rc::Rc::new(::std::cell::Cell::new(None));
+    };
+
+    let adaptor_var_names = || {
+        (0 .. arms.len())
+            .map(|i| Ident::new(&format!("adaptor{}", i), Span::mixed_site()))
+    };
+
+    let adaptor_decls = adaptor_var_names()
+        .zip(future_var_names())
+        .zip(&arms)
+        .map(|((adaptor, future), arm)| {
+            let pat = &arm.pattern;
+            let final_output = &arm.output;
+            quote! {
+                let #adaptor = {
+                    let #output_var_name = #output_var_name.clone();
+                    async move {
+                        let output_val = #future.await;
+                        if #output_var_name.get().is_none() {
+                            let #pat = output_val;
+                            #output_var_name.set(Some(#final_output));
+                        }
+                        Ok(::webio::wasm_bindgen::JsValue::UNDEFINED)
+                    }
+                };
+            }
+        });
+
+    let promise_var_names = || {
+        (0 .. arms.len())
+            .map(|i| Ident::new(&format!("promise{}", i), Span::mixed_site()))
+    };
+
+    let promise_decls = promise_var_names().zip(adaptor_var_names()).map(
+        |(promise, adaptor)| {
+            quote! {
+                let #promise = ::webio::wasm_bindgen::JsValue::from(
+                    ::webio::wasm_bindgen_futures::future_to_promise(#adaptor)
+                );
+            }
+        },
+    );
+
+    let promise_var_names_iter = promise_var_names();
+
+    let expanded = quote! {
+        {
+            #(#future_decls)*
+            #output_decl
+            #(#adaptor_decls)*
+            #(#promise_decls)*
+            let mut promise_list = ::webio::js_sys::Array::new();
+            promise_list.extend([#(#promise_var_names_iter),*]);
+            let final_promise = ::webio::js_sys::Promise::any(&promise_list);
+            ::webio::wasm_bindgen_futures::JsFuture::from(final_promise)
+                .await
+                .unwrap();
+            #output_var_name.take().unwrap()
+        }
+    };
+
+    expanded.into_token_stream().into()
+}
+
 struct ConsoleInput {
     method: Ident,
     arguments: Vec<Expr>,
@@ -145,23 +317,31 @@ impl Parse for ConsoleInput {
     fn parse(input: ParseStream) -> syn::Result<Self> {
         let method = input.parse()?;
         let mut arguments = Vec::new();
-        if input.peek(Semi) {
-            input.parse::<Semi>()?;
+        if input.peek(token::Semi) {
+            input.parse::<token::Semi>()?;
             arguments.extend(
-                input.parse_terminated::<Expr, Comma>(Expr::parse)?.into_iter(),
+                input
+                    .parse_terminated::<Expr, token::Comma>(Expr::parse)?
+                    .into_iter(),
             );
         }
         Ok(Self { method, arguments })
     }
 }
 
-/// Prints to the JavaScript/browser/node console using a given method. Syntax:
+/// Prints to the JavaScript/browser/node console using a given method.
+///
+/// Syntax:
+///
 /// ```ignore
-/// console!($method; $($arguments),*)
+/// console!($method; argument0, argument1, argument2, ..., argument_n)
 /// ```
+///
 /// Each argument is converted into a `JsValue` using `Into`.
 ///
 /// # Examples
+///
+/// ## Log Different ARgument Types
 ///
 /// ```ignore
 /// use webio::console;
@@ -207,7 +387,7 @@ pub fn console(raw_input: TokenStream) -> TokenStream {
     expanded.into()
 }
 
-/// This macro converts an asynchronous main function into a synchronous one,
+/// This macro converts an asynchronous main function into a synchronous one
 /// that can actually be an entry point, and that invokes the asynchronous code.
 /// Under the hood, the asynchronous code is detached from the current call.
 ///
@@ -262,6 +442,12 @@ pub fn main(raw_attribute: TokenStream, raw_input: TokenStream) -> TokenStream {
             "webio::main does not support ABIs",
         ));
     }
+    if input.sig.unsafety.is_some() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::main cannot be unsafe",
+        ));
+    }
 
     match errors {
         Some(stored) => stored.into_compile_error().into(),
@@ -271,12 +457,96 @@ pub fn main(raw_attribute: TokenStream, raw_input: TokenStream) -> TokenStream {
             let ident = input.sig.ident;
             let body = input.block;
             let attrs = input.attrs;
-            let unsafety = input.sig.unsafety;
-            let abi = input.sig.abi;
             let expanded = quote! {
                 #(#attrs)*
                 #[::webio::wasm_bindgen::prelude::wasm_bindgen(start)]
-                #visibility #unsafety #abi #fn_token #ident() {
+                #visibility #fn_token #ident() {
+                    ::webio::task::detach(async {
+                        let (): () = #body;
+                    });
+                }
+            };
+            expanded.into_token_stream().into()
+        },
+    }
+}
+
+/// This macro converts an asynchronous test function into a synchronous one
+/// that can actually be tested by `wasm_bindgen_test`, and that invokes the
+/// asynchronous code. Under the hood, the asynchronous code is detached from
+/// the current call.
+///
+/// # Examples
+///
+/// ## Test With Timeout
+///
+/// ```ignore
+/// use webio::time::timeout;
+/// use std::time::Duration;
+///
+/// #[webio::test]
+/// async fn my_test() {
+///     timeout(Duration::from_millis(200)).await;
+///     assert!(true);
+/// }
+/// ```
+#[proc_macro_attribute]
+pub fn test(raw_attribute: TokenStream, raw_input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(raw_input as ItemFn);
+    let mut errors: Option<syn::Error> = None;
+    let mut append_error = |error| match errors.as_mut() {
+        Some(stored) => stored.combine(error),
+        None => errors = Some(error),
+    };
+    if !raw_attribute.is_empty() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::test attribute must not receive arguments",
+        ));
+    }
+    if input.sig.asyncness.is_none() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::test must be an asynchronous function with async syntax",
+        ));
+    }
+    if input.sig.constness.is_some() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::test cannot be const",
+        ));
+    }
+    if input.sig.inputs.len() > 0 || input.sig.variadic.is_some() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::test function cannot receive parameters",
+        ));
+    }
+    if input.sig.abi.is_some() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::test does not support ABIs",
+        ));
+    }
+    if input.sig.unsafety.is_some() {
+        append_error(syn::Error::new(
+            Span::call_site(),
+            "webio::test cannot be unsafe",
+        ));
+    }
+
+    match errors {
+        Some(stored) => stored.into_compile_error().into(),
+        None => {
+            let visibility = input.vis;
+            let fn_token = input.sig.fn_token;
+            let ident = input.sig.ident;
+            let body = input.block;
+            let attrs = input.attrs;
+            let expanded = quote! {
+                #(#attrs)*
+                #[::webio::wasm_bindgen_test::wasm_bindgen_test]
+                #visibility #fn_token #ident() {
                     ::webio::task::detach(async {
                         let (): () = #body;
                     });